code style

+++ FileSplitter.pas
@@ -100,14 +100,12 @@ begin
 	self.ExternalProgressProc := ExternalProgressProc;
 	self.ExternalPluginNr := PluginNr;
 	self.totalPartsCount := self.Splitted.size div self.PartSize;
-	if (self.Splitted.size mod self.PartSize) <> 0 then
-		inc(self.totalPartsCount);
+	if (self.Splitted.size mod self.PartSize) <> 0 then inc(self.totalPartsCount);
 
 	self.Splitted.path := IncludeTrailingBackslash(GetTmpDir + GetCurrentProcessId.ToString());
 	if not DirectoryExists(self.Splitted.path) then
 	begin
-		if not CreateDir(self.Splitted.path) then
-			Raise Exception.CreateFmt('Can''t create temp directory : ''%s''', [self.Splitted.path]);
+		if not CreateDir(self.Splitted.path) then Raise Exception.CreateFmt('Can''t create temp directory : ''%s''', [self.Splitted.path]);
 	end;
 	//self.Splitted.crc32 := self.getCRC32File(filename);
 end;
@@ -118,11 +116,9 @@ var
 begin
 	for SplittedPartIndex := 0 to Length(self.SplitResult.parts) - 1 do
 	begin
-		if FileExists(self.SplitResult.parts[SplittedPartIndex].filename) then
-			DeleteFileW(PWideChar(self.SplitResult.parts[SplittedPartIndex].filename));
+		if FileExists(self.SplitResult.parts[SplittedPartIndex].filename) then DeleteFileW(PWideChar(self.SplitResult.parts[SplittedPartIndex].filename));
 	end;
-	if FileExists(self.Splitted.path + ChangeFileExt(ExtractFileName(self.Splitted.filename), '.crc')) then
-		DeleteFileW(PWideChar(self.Splitted.path + ChangeFileExt(ExtractFileName(self.Splitted.filename), '.crc')));
+	if FileExists(self.Splitted.path + ChangeFileExt(ExtractFileName(self.Splitted.filename), '.crc')) then DeleteFileW(PWideChar(self.Splitted.path + ChangeFileExt(ExtractFileName(self.Splitted.filename), '.crc')));
 	inherited;
 end;
 
@@ -157,8 +153,7 @@ begin
 	except
 		exit(FS_FILE_READERROR);
 	end;
-	if partNumber >= self.totalPartsCount then
-		exit(FS_FILE_NOTSUPPORTED);
+	if partNumber >= self.totalPartsCount then exit(FS_FILE_NOTSUPPORTED);
 
 	FStream.Position := self.PartSize * partNumber;
 
@@ -166,8 +161,7 @@ begin
 	begin
 		PartSize := self.Splitted.size - (self.PartSize * partNumber);
 	end
-	else
-		PartSize := self.PartSize;
+	else PartSize := self.PartSize;
 	partStream.CopyFrom(FStream, PartSize);
 	result := FS_FILE_OK;
 end;
@@ -212,8 +206,7 @@ begin
 		begin
 			PartSize := self.Splitted.size - (self.PartSize * partsCount);
 		end
-		else
-			PartSize := self.PartSize;
+		else PartSize := self.PartSize;
 		try
 			WStream.CopyFrom(FStream, PartSize);
 		except
